# backend/doc/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "High Street Gym API",
    "version": "1.0.0",
    "description": "JSON REST API for interacting with the High Street Gym Backend"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-auth-key",
        "description": "API key for authentication"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "An error occurred"
          }
        }
      },
...
      "Session": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "The unique identifier of the session",
            "example": 1
          },
          "activityId": {
            "type": "number",
            "description": "The ID of the activity associated with the session",
            "example": 1
          },
          "trainerId": {
            "type": "number",
            "description": "The ID of the trainer leading the session",
            "example": 2
          },
          "locationId": {
            "type": "number",
            "description": "The ID of the location where the session is held",
            "example": 3
          },
          "sessionDate": {
            "type": "string",
            "format": "date",
            "description": "The date of the session (YYYY-MM-DD)",
            "example": "2025-11-01"
          },
          "sessionTime": {
            "type": "string",
            "description": "The time of the session (HH:MM:SS)",
            "example": "10:00:00"
          },
          "deleted": {
            "type": "number",
            "description": "Deletion status (0 = active, 1 = deleted)",
            "example": 0
          }
        }
      },
      "SessionInput": {
        "type": "object",
        "required": [
          "activityId",
          "locationId",
          "sessionDate",
          "sessionTime"
        ],
        "properties": {
          "activityId": {
            "type": "number",
            "description": "The ID of the activity associated with the session",
            "example": 1
          },
          "locationId": {
            "type": "number",
            "description": "The ID of the location where the session is held",
            "example": 3
          },
          "sessionDate": {
            "type": "string",
            "format": "date",
            "description": "The date of the session (YYYY-MM-DD)",
            "example": "2025-11-01"
          },
          "sessionTime": {
            "type": "string",
            "description": "The time of the session (HH:MM:SS)",
            "example": "10:00:00"
          }
        },
        "example": {
          "activityId": 1,
          "locationId": 3,
          "sessionDate": "2025-11-01",
          "sessionTime": "10:00:00"
        }
      },
...
    },
    "responses": {
...
      "BadRequest": {
        "description": "Invalid request body or bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or invalid credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "Access forbidden - insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "SuccessMessage": {
        "description": "Success response with a message",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "description": "Success message",
                  "example": "Operation successful"
                }
              }
            }
          }
        }
      },
...
      "SessionCreated": {
        "description": "Session created successfully",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Session"
            }
          }
        }
      },
      "SessionsList": {
        "description": "List of sessions",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Session"
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/sessions": {
      "get": {
        "summary": "Get all sessions",
        "tags": [
          "Sessions"
        ],
        "description": "Retrieve a list of all active sessions, sorted by date and time",
        "responses": {
          "200": {
            "$ref": "#/components/responses/SessionsList"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Create a new session",
        "tags": [
          "Sessions"
        ],
        "description": "Create a session for the authenticated trainer. Trainers can only create sessions for themselves.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Session request containing activity, location, date, and time",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/SessionCreated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/sessions/self": {
      "get": {
        "summary": "Get authenticated trainer's sessions",
        "tags": [
          "Sessions"
        ],
        "description": "Retrieve a list of all active sessions for the authenticated trainer, sorted by date and time",
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SessionsList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/sessions/{id}": {
      "delete": {
        "summary": "Cancel a session",
        "tags": [
          "Sessions"
        ],
        "description": "Cancel a session by ID. Only the trainer who owns the session can cancel it.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The session ID"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessMessage"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/sessions/export/xml/weekly": {
      "get": {
        "summary": "Export trainer's weekly sessions as XML",
        "tags": [
          "Sessions"
        ],
        "description": "Export the authenticated trainer's sessions as XML for calendar import. Supports optional date range filtering via query parameters.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Start date for filtering sessions (YYYY-MM-DD format)"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "End date for filtering sessions (YYYY-MM-DD format)"
          }
        ],
        "responses": {
          "200": {
            "description": "XML file containing trainer's weekly sessions",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string",
                  "example": "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<calendar>\n  <header>\n    <title>Sessions - John Doe</title>\n    <trainer>\n      <name>John Doe</name>\n      <email>trainer@hsg.com</email>\n    </trainer>\n  </header>\n  <sessions>...</sessions>\n</calendar>\n"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
...
}